Refactor to have a dedicated file for relation_column_stats
