+++ input/postgres/relation_stats.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 
 	"github.com/pganalyze/collector/state"
-	"github.com/pganalyze/collector/util"
 )
 
 const relationStatsSQLInsertsSinceVacuumFieldPg13 string = "s.n_ins_since_vacuum"
@@ -142,19 +141,6 @@ SELECT relid,
   FROM locked_relids
 `
 
-const columnStatsSQL = `
-SELECT schemaname, tablename, attname, inherited, null_frac, avg_width, n_distinct, correlation
-  FROM %s
- WHERE schemaname NOT IN ('pg_catalog', 'information_schema')
-`
-
-const columnStatsHelperSQL = `
-SELECT 1 AS enabled
-  FROM pg_catalog.pg_proc p
-  JOIN pg_catalog.pg_namespace n ON (p.pronamespace = n.oid)
- WHERE n.nspname = 'pganalyze' AND p.proname = 'get_column_stats'
-`
-
 func GetRelationStats(ctx context.Context, db *sql.DB, postgresVersion state.PostgresVersion, server *state.Server) (relStats state.PostgresRelationStatsMap, err error) {
 	var insertsSinceVacuumField string
 
@@ -253,56 +239,4 @@ func GetIndexStats(ctx context.Context, db *sql.DB, postgresVersion state.Postgr
 	indexStats, err = handleIndexStatsExt(ctx, db, indexStats, postgresVersion, server)
 
 	return
-}
-
-func GetColumnStats(ctx context.Context, logger *util.Logger, db *sql.DB, globalCollectionOpts state.CollectionOpts, systemType string, dbName string) (columnStats state.PostgresColumnStatsMap, err error) {
-	var sourceTable string
-
-	helperExists := false
-	db.QueryRow(QueryMarkerSQL + columnStatsHelperSQL).Scan(&helperExists)
-
-	if helperExists {
-		logger.PrintVerbose("Found pganalyze.get_column_stats() stats helper")
-		sourceTable = "pganalyze.get_column_stats()"
-	} else {
-		if systemType != "heroku" && !connectedAsSuperUser(ctx, db, systemType) && globalCollectionOpts.TestRun {
-			logger.PrintInfo("Warning: Limited access to table column statistics detected in database %s. Please set up"+
-				" the monitoring helper function pganalyze.get_column_stats (https://github.com/pganalyze/collector#setting-up-a-restricted-monitoring-user)"+
-				" or connect as superuser, to get column statistics for all tables.", dbName)
-		}
-		sourceTable = "pg_catalog.pg_stats"
-	}
-
-	stmt, err := db.PrepareContext(ctx, QueryMarkerSQL+fmt.Sprintf(columnStatsSQL, sourceTable))
-	if err != nil {
-		return nil, err
-	}
-	defer stmt.Close()
-
-	rows, err := stmt.QueryContext(ctx)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var statsMap = make(state.PostgresColumnStatsMap)
-
-	for rows.Next() {
-		var s state.PostgresColumnStats
-
-		err := rows.Scan(
-			&s.SchemaName, &s.TableName, &s.ColumnName, &s.Inherited, &s.NullFrac, &s.AvgWidth, &s.NDistinct, &s.Correlation)
-		if err != nil {
-			return nil, err
-		}
-
-		key := state.PostgresColumnStatsKey{SchemaName: s.SchemaName, TableName: s.TableName, ColumnName: s.ColumnName}
-		statsMap[key] = append(statsMap[key], s)
-	}
-
-	if err = rows.Err(); err != nil {
-		return nil, err
-	}
-
-	return statsMap, nil
 }
